request.user

+++ core/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, get_object_or_404
 from core.models import Articulos
 from core.forms import ArticulosForm
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 
 
 
@@ -25,7 +26,7 @@ def crea_articulos(request):
            articulo = Articulos(titulo=titulo, subtitulo=subtitulo, cuerpo=cuerpo, autor=autor, imagen=imagen)
            articulo.save()
            id = articulo.id
-           return render(request, 'core/mostrar_articulos.html', {'articulo.id':id, 'articulo':articulo}) #si sale todo bien nos manda a la lista de articulos
+           return render(request, 'core/mostrar_postcrear.html', {"mensaje": "Recomendacion agregada correctamente",'articulo':articulo}) #si sale todo bien nos manda a la lista de articulos
     else:
         articulos_form = ArticulosForm()
     return render (request, 'core/crear_articulos.html', {"form": articulos_form})
@@ -54,22 +55,35 @@ def edita_articulos(request, id_articulo):
             articulo.subtitulo = data["subtitulo"]
             articulo.cuerpo= data["cuerpo"]
             articulo.autor = data["autor"]
-            articulo.imagen = data["imagen"]
+            if 'imagen' in request.FILES:
+             articulo.imagen = data["imagen"]
 
-            articulo.save()
-            return render(request, 'core/mostrar_articulos.html', {'articulo':articulo})
+#verificamos si el usuario que intenta editar es el que creó el artículo
+            if articulo.usuario == request.user:
+               articulo.save()
+               return render(request, 'core/mostrar_posteditar.html', {'articulo':articulo})
+            else:
+                return render(request, 'core/autorización.html')
+        
     else: #metodo get
-        articulo_form=ArticulosForm(initial={'titulo':articulo.titulo, 'subtitulo':articulo.subtitulo, 'cuerpo':articulo.cuerpo,'autor':articulo.autor, 'imagen':articulo.imagen})
-    return render (request, 'core/editar_articulos.html', {'articulo_form':articulo_form, 'articulo':articulo })
+        if articulo.usuario == request.user or request.user.is_staff:
+           articulo_form=ArticulosForm(initial={'titulo':articulo.titulo, 'subtitulo':articulo.subtitulo, 'cuerpo':articulo.cuerpo,'autor':articulo.autor, 'imagen':articulo.imagen})
+           return render (request, 'core/editar_articulos.html', {'articulo_form':articulo_form})
+        else:
+             return render(request, 'core/autorización.html')
 
 @login_required
 def elimina_articulos(request, id_articulo): 
 
     articulo = Articulos.objects.get(id=id_articulo)
     name = articulo.titulo
-    articulo.delete()
-    
-    return render (request, 'core/eliminar_articulos.html', {"articulo_eliminado": name} )
+
+    if articulo.usuario == request.user:          
+      articulo.delete()
+      return render (request, 'core/eliminar_articulos.html', {"articulo_eliminado": name} )
+    else:
+        return render(request, 'core/autorización.html')
+       
 
 def acerca_de_mi(request):
     return render (request, 'core/about.html')
